adding test fixes

+++ tests/unit/test_dask_nvt.py
@@ -17,7 +17,7 @@
 import glob
 import math
 import os
-
+import numpy as np
 import cudf
 import dask_cudf
 import pytest
@@ -165,7 +165,7 @@ def test_dask_groupby_stats(client, tmpdir, datasets, part_mem_fraction):
         result[["name-cat", "name-cat_count"]]
         .drop_duplicates()
         .sort_values("name-cat")["name-cat_count"],
-        df0.groupby("name-cat").agg({"x": "count"})["x"],
+        df0.groupby("name-cat").agg({"x": "count"})["x"].astype(np.int64),
         check_index=False,
         check_dtype=False,  # May get int64 vs int32
         check_names=False,
